test: 리팩토링으로 인한 테스트 재구성
